ESWA_Dynamic_Ensemble_Trading_GitHub/requirements.txt
# ESWA 동적 앙상블 강화학습 거래 시스템 의존성 패키지
# 핵심 강화학습 및 딥러닝 라이브러리
torch>=1.12.0
torchvision>=0.13.0
stable-baselines3>=1.6.0
gymnasium>=0.28.0

# 멀티모달 특징 추출
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 기술적 지표 계산
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
ta>=0.10.0  # Technical Analysis library

# 시장 체제 분류
xgboost>=1.6.0
scikit-learn>=1.1.0

# 뉴스 감정분석
transformers>=4.20.0
requests>=2.28.0
beautifulsoup4>=4.11.0
textblob>=0.17.0

# 거래 및 금융 데이터
ccxt>=2.0.0  # Cryptocurrency exchange library
yfinance>=0.1.70
alpha-vantage>=2.3.0

# 데이터 시각화 및 분석
plotly>=5.10.0
bokeh>=2.4.0
jupyter>=1.0.0
ipykernel>=6.0.0

# 설정 및 로깅
pyyaml>=6.0
python-dotenv>=0.19.0
loguru>=0.6.0

# 테스트 및 개발 도구
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# 성능 최적화
numba>=0.56.0
joblib>=1.1.0

# 추가 유틸리티
tqdm>=4.64.0
colorama>=0.4.5
rich>=12.0.0
